# car-rental/src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Car Rental CRUD API
  description: API for managing cars
  version: 1.0.0

servers:
  - url: http://localhost:8080/api/v1
  - url: http://localhost:8080/api/v2

security:
  - api_key: [ ]

components:
  schemas:
    CarDTO:
      type: object
      required:
        - id
        - brand
        - model
        - available
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the car
        brand:
          type: string
          description: Brand of the car
        model:
          type: string
          description: Model of the car
        available:
          type: boolean
          description: Availability of the car

paths:
  /api/v1/cars:
    post:
      summary: Create a new car
      operationId: createCar
      requestBody:
        description: Car object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarDTO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarDTO'
    put:
      summary: Update a car by ID
      operationId: updateCarById
      requestBody:
        description: Car object to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarDTO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarDTO'
        '404':
          description: Car not found
  /api/v1/cars/{id}:
    delete:
      summary: Delete a car by ID
      operationId: deleteCarById
      parameters:
        - name: id
          in: path
          description: ID of the car to delete
          required: true
          schema:
            type: integer
            format: int64
  /api/v1/cars/available:
    get:
      summary: List all available cars by model or brand
      operationId: listAvailableCars
      parameters:
        - name: modelName
          in: query
          description: Filter by model
          schema:
            type: string
        - name: brandName
          in: query
          description: Filter by brand
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarDTO'